synq: widen _layer_id from uint8_t to uint32_t to fix overflow at 256 layers (#128)
When a single statement produces more than 255 expansion layers (e.g.
256+ macro calls in a VALUES table), the uint8_t _layer_id wraps to 0.
This makes the span machinery believe macro-produced spans came from
the original source, causing node_text / span_expanded_text to return
wrong bytes.

Widen _layer_id, parent_layer_id, and layer_id to uint32_t across all
structs (SyntaqliteTextSpan, SyntaqliteParserToken, SynqParseToken,
SynqExpansionLayer) and their Rust FFI mirrors. Remove the truncating
(uint8_t) casts in parser.c and parser_macros.c. Also widen the span
flags field to uint32_t for consistency and to eliminate padding.

Fixes #128.

+++ syntaqlite-syntax/csrc/parser_internal.h
@@ -84,7 +84,7 @@ typedef struct SynqExpansionLayer {
   SynqArgSegment* arg_segments;
   uint32_t arg_segment_count;
 
-  uint8_t parent_layer_id;  // Layer containing the call (0 = source).
+  uint32_t parent_layer_id;  // Layer containing the call (0 = source).
 } SynqExpansionLayer;
 
 typedef SYNQ_VEC(SynqExpansionLayer) SynqExpansionLayerVec;
